requirements.txt
# Optional runtime packages. AgentShield falls back gracefully when unavailable.
numpy>=1.26.0
scikit-learn>=1.5.0
joblib>=1.4.0
PyYAML>=6.0.0
streamlit>=1.37.0
pandas>=2.2.0
